+++ projects/swimlane/ngx-ui/src/lib/components/input/input.component.ts
@@ -365,6 +365,14 @@ export class InputComponent implements AfterViewInit, OnDestroy, ControlValueAcc
     this.disabled = coerceBooleanProperty(isDisabled);
   }
 
+  unlock(): void {
+    if (this.type === InputTypes.password) {
+      this.value = '';
+    }
+    this.disabled = false;
+    this.updateInputType();
+  }
+
   incrementValue(event: MouseEvent): void {
     if (this.disabled) return;